build(gradle): migrate to Gradle 9 convention plugins and centralize versions
- Move build logic into included build  with precompiled convention plugins
- Apply plugins: cryptad.java-kotlin-conventions, cryptad.spotless, cryptad.versioning, cryptad.buildjar, cryptad.distribution
- Centralize plugin/artifact versions in root version catalog; remove hardcoded versions
- Upgrade Kotlin to 2.2.10
- Remove unused publishing convention plugin
- Keep strict dependency verification; update verification metadata

No CHANGELOG updates.
