+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-kotlin = "2.2.0"
+kotlin = "2.2.10"
 bcprov = "2.73.7"
 jna = "5.17.0"
 commonsCompress = "1.28.0"
@@ -10,6 +10,8 @@ junit4 = "4.13.2"
 mockito = "5.18.0"
 hamcrest = "3.0"
 objenesis = "3.4"
+picocli = "4.7.7"
+spotless = "7.2.1"
 
 [libraries]
 bcprov = { group = "org.bouncycastle", name = "bcprov-lts8on", version.ref = "bcprov" }
@@ -24,6 +26,8 @@ junit4 = { group = "junit", name = "junit", version.ref = "junit4" }
 mockitoCore = { group = "org.mockito", name = "mockito-core", version.ref = "mockito" }
 hamcrest = { group = "org.hamcrest", name = "hamcrest", version.ref = "hamcrest" }
 objenesis = { group = "org.objenesis", name = "objenesis", version.ref = "objenesis" }
+picocli = { group = "info.picocli", name = "picocli", version.ref = "picocli" }
 
 [plugins]
 kotlinJvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
+spotless = { id = "com.diffplug.spotless", version.ref = "spotless" }
