+++ gradle/verification-metadata.xml
@@ -29,6 +29,7 @@
          <trusted-key id="144219B604DDE6A2383FA0EE9C6404EBCE3E4464" group="com.github.zafarkhaja" name="java-semver" version="0.9.0"/>
          <trusted-key id="147B691A19097624902F4EA9689CBE64F4BC997F" group="org.mockito" name="mockito-core" version="5.18.0"/>
          <trusted-key id="19BEAB2D799C020F17C69126B16698A4ADF4D638" group="org.checkerframework" name="checker-qual" version="3.37.0"/>
+         <trusted-key id="1BD97A6A154E7810EE0BC832E2F38302C8075E3D" group="org.gradle.kotlin" name="gradle-kotlin-dsl-plugins" version="6.2.0"/>
          <trusted-key id="1D2C7EF8ADA0F794B58C7C63436902AF59EDF60E">
             <trusting group="dev.equo.ide" name="solstice" version="1.7.5"/>
             <trusting group="dev.equo.ide" name="solstice" version="1.8.1"/>
@@ -901,6 +902,11 @@
       <component group="org.freenetproject" name="freenet-ext" version="29">
          <artifact name="freenet-ext-29.jar">
             <sha256 value="32f2b3d6beedf54137ea2f9a3ebef67666d769f0966b08cd17fd7db59ba4d79f" origin="Generated by Gradle because artifact wasn't signed" reason="Artifact is not signed"/>
+         </artifact>
+      </component>
+      <component group="org.gradle.kotlin.kotlin-dsl" name="org.gradle.kotlin.kotlin-dsl.gradle.plugin" version="6.2.0">
+         <artifact name="org.gradle.kotlin.kotlin-dsl.gradle.plugin-6.2.0.pom">
+            <sha256 value="b4851c9aeb6e79aa17a95288315c54fa2b3517c4cdbbc2da0b7f1b710c1d2d71" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="org.hamcrest" name="hamcrest" version="3.0">
