+++ settings.gradle.kts
@@ -11,3 +11,6 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "cryptad"
+
+// Gradle 9: Use an included build for convention plugins instead of buildSrc
+includeBuild("build-logic")
